+++ src/common.ts
@@ -1,8 +1,0 @@
-export type Address = {
-   street: string
-   city: string
-   zip: string
-   country: string
-}
-
-export type Currency = "EUR" | "USD" | "CZK"