+++ src/index.ts
@@ -3,11 +3,11 @@ import { Payer } from "./payment/Payer"
 import { Payment } from "./payment/Payment"
 import { PaymentItem } from "./payment/PaymentItem"
 import { Price } from "./payment/Price"
-import { Invoice, InvoiceData } from "./invoicing/Invoice"
+import { Invoice } from "./invoicing/Invoice"
 import { TaxRate } from "./payment/TaxRate"
+import { InvoiceMetadata } from "./invoicing/invoiceMetadata"
 
 import { StripeClient } from "./stripe/Client"
-import { StripeWebhook } from "./stripe/WebhookHandler"
 import { StripeCheckoutSession } from "./stripe/CheckoutSession"
 
 export {
@@ -17,9 +17,8 @@ export {
    PaymentItem,
    Price,
    Invoice,
-   InvoiceData,
    StripeClient,
    TaxRate,
-   StripeWebhook,
-   StripeCheckoutSession
+   StripeCheckoutSession,
+   InvoiceMetadata,
 }