+++ src/invoicing/Invoice.ts
@@ -1,104 +1,30 @@
 import { Result } from "@flagg2/result"
-import { Payee } from "../payment/Payee"
-import { Payer } from "../payment/Payer"
 import z from "zod"
-import { Payment } from "../payment/Payment"
 import { PaymentItem } from "../payment/PaymentItem"
 import Decimal from "decimal.js"
 import { briskFetch } from "../utils/briskFetch"
+import { InvoiceMetadata } from "./invoiceMetadata"
+import { positiveInteger } from "../common/decimal"
+import { Currency } from "../common/currency"
+import { Payer } from "../payment/Payer"
+import { Payee } from "../payment/Payee"
 
-type Invoice = {
-   invoiceData: InvoiceData
-   payment: Payment
-   payee: Payee
-   payer: Payer
-}
+type Invoice = z.infer<typeof schema>
 
-type InvoiceData = {
-   createdDate: Date
-   deliveryDate: Date
-   dueDate: Date
-   number: string
-   numberPrefix: string
-   status: "unpaid" | "partiallyPaid" | "paid"
-   paymentMethod: "cash" | "card" | "transfer"
-   logoUrl?: string
-   note?: string
-   variableSymbol?: number
-}
-
-const companyInfoSchema = z.object({
-   businessId: z.string().nonempty(),
-   taxId: z.string().optional(),
-   vatId: z.string().optional(),
-})
-
-const addressSchema = z.object({
-   street: z.string().nonempty(),
-   city: z.string().nonempty(),
-   zip: z.string().nonempty(),
-   country: z.string().nonempty(),
-})
-
-const paymentItemSchema = z.object({
-   name: z.string().nonempty(),
-   priceWithoutTax: z.coerce
-      .number()
-      .positive()
-      .transform((value) => new Decimal(value)),
-   taxRate: z.coerce
-      .number()
-      .positive()
-      .transform((value) => new Decimal(value)),
-   description: z.string().optional(),
-   imageUrl: z.string().optional(),
-})
-
-const mapLikeItemsSchema = z.map(
-   paymentItemSchema,
-   z.coerce
-      .number()
-      .int()
-      .positive()
-      .transform((value) => new Decimal(value)),
-)
+const mapLikeItemsSchema = z.map(PaymentItem.getSchema(), positiveInteger)
 const arrayLikeItemsSchema = z.array(
    z.object({
-      item: paymentItemSchema,
-      quantity: z.coerce
-         .number()
-         .int()
-         .positive()
-         .transform((value) => new Decimal(value)),
+      item: PaymentItem.getSchema(),
+      quantity: positiveInteger,
    }),
 )
 
 const schema = z.object({
-   invoiceData: z.object({
-      createdDate: z.coerce.date(),
-      deliveryDate: z.coerce.date(),
-      dueDate: z.coerce.date(),
-      number: z.string().nonempty(),
-      numberPrefix: z.string().nonempty(),
-      status: z.enum(["unpaid", "partiallyPaid", "paid"]),
-      paymentMethod: z.enum(["cash", "card", "transfer"]),
-      logoUrl: z.string().optional(),
-      note: z.string().optional(),
-      variableSymbol: z.coerce.number().int().positive().optional(),
-   }),
-   payee: z.object({
-      name: z.string().nonempty(),
-      billingInfo: addressSchema,
-      companyInfo: companyInfoSchema,
-   }),
-   payer: z.object({
-      name: z.string().nonempty(),
-      billingInfo: addressSchema,
-      shippingInfo: addressSchema.optional(),
-      companyInfo: companyInfoSchema.optional(),
-   }),
+   meta: InvoiceMetadata.getSchema(),
+   payee: Payee.getSchema(),
+   payer: Payer.getSchema(),
    payment: z.object({
-      currency: z.enum(["CZK", "EUR", "USD"]),
+      currency: Currency.getSchema(),
       items: z
          .union([mapLikeItemsSchema, arrayLikeItemsSchema])
          .transform((items) => {
@@ -211,11 +137,23 @@ async function createPdf(
    return result.map((res) => ({ base64: res.data }))
 }
 
+/**
+ * Creates an invoice.
+ *
+ * @param invoice The invoice to create.
+ * @returns The created invoice.
+ */
+
+function create(invoice: Invoice) {
+   return invoice
+}
+
 const Invoice = {
    parseUnknown,
    getSchema,
    createPdf,
+   create,
    toPlainObject,
 }
 
-export { Invoice, InvoiceData }
+export { Invoice }